Recursively add reactions when multiple emoji are found

+++ src/listeners/EmojiListener.ts
@@ -2,27 +2,44 @@ import { Message } from 'discord.js';
 import { IListener } from '../interfaces/Listener';
 import Jarvis from '../client/Jarvis';
 import * as e from 'node-emoji';
+import * as runes from 'runes';
 
 export default class EmojiListener implements IListener
 {
 	public constructor() {}
 
 	/**
-	 * Listen for `+:emoji:` type messages to effectively
-	 * implement the feature for mobile client use since
-	 * I don't currently have that feature on android
+	 * Listen for `+:emoji:` messages and recursively react with
+	 * all the given emojis and custom emojis to the most recent
+	 * message before the command
 	 */
-	public async process(jarvis: Jarvis, message: Message): Promise<any>
+	public async process(jarvis: Jarvis, message: Message, id?: string): Promise<any>
 	{
 		if (message.author.id !== jarvis.user.id) return;
 		if (!message.content.startsWith('+')) return;
-		message.delete();
-		const m: string = message.content.slice(1).trim();
+		if (!id) message.delete();
+		let m: string = message.content.slice(1).trim();
+		const findCustom: RegExp = /\<:[^:]+:\d+\>/g;
 		const parseCustom: RegExp = /\<:[^:]+:(\d+)\>/;
-		const emoji: string = parseCustom.test(m) ? m.match(parseCustom)[1] : e.which(m);
-		const toReact: Message = (await message.channel
-			.fetchMessages({ limit: 1, before: message.id })).first();
+		const customEmojis: string[] = m.match(findCustom);
+		m = m.replace(findCustom, '');
+
+		let emojis: string[] = runes(m).map((a: string) => a.trim()).filter((a: string) => a !== '');
+		let emoji: string;
+
+		if (emojis && emojis.length > 0) emoji = e.which(emojis.shift());
+		else if (customEmojis && customEmojis.length > 0) emoji = customEmojis.shift().match(parseCustom)[1];
+
+		const toReact: Message = id ? await message.channel.fetchMessage(id)
+			: (await message.channel.fetchMessages({ limit: 1, before: message.id })).first();
 		if (!emoji || !toReact) return;
-		return toReact.react(jarvis.emojis.get(emoji) || e.get(emoji));
+
+		toReact.react(jarvis.emojis.get(emoji) || e.get(emoji));
+		emojis = emojis.concat(customEmojis);
+		if (emojis && emojis.length > 0)
+		{
+			message.content = `+${emojis.join('')}`;
+			this.process(jarvis, message, toReact.id);
+		}
 	}
 }
